+++ backend/api/controllers/leaderboards.ts
@@ -7,12 +7,14 @@ class LeaderboardsController {
     const { language, mode, mode2, skip, limit = 50 } = req.query;
     const { uid } = req.ctx.decodedToken;
 
+    const queryLimit = Math.min(parseInt(limit as string, 10), 50);
+
     const leaderboard = await LeaderboardsDAO.get(
       mode,
       mode2,
       language,
-      parseInt(skip as string),
-      parseInt(limit as string)
+      parseInt(skip as string, 10),
+      queryLimit
     );
 
     const normalizedLeaderboard = _.map(leaderboard, (entry) => {
