+++ backend/package.json
@@ -35,6 +35,7 @@
     "simple-git": "2.45.1",
     "string-similarity": "4.0.4",
     "swagger-stats": "0.99.2",
+    "swagger-ui-express": "4.3.0",
     "ua-parser-js": "0.7.28",
     "uuid": "8.3.2"
   },
@@ -46,6 +47,7 @@
     "@types/node": "17.0.18",
     "@types/node-fetch": "2.6.1",
     "@types/swagger-stats": "0.95.4",
+    "@types/swagger-ui-express": "4.1.3",
     "@types/ua-parser-js": "0.7.36",
     "@types/uuid": "8.3.4"
   }
