+++ backend/tsconfig.json
@@ -13,5 +13,5 @@
     "esModuleInterop": true,
     "strictNullChecks": true
   },
-  "exclude": ["node_modules", "build"]
+  "exclude": ["node_modules", "build", "worker.js"]
 }
