+++ frontend/static/index.html
@@ -4159,10 +4159,18 @@
                   </div>
                 </div>
               </div>
-              <!-- <div class="section apeKeys needsAccount">
+              <div class="section apeKeys needsAccount">
                 <h1>ape keys</h1>
                 <div class="text">
                   Generate Ape Keys to access certain API endpoints.
+                  Documentation is available
+                  <a
+                    href="https://api.monkeytype.com/documentation"
+                    target="_blank"
+                  >
+                    here.
+                  </a>
+                  More endpoints will be added in the future.
                 </div>
                 <div class="buttons">
                   <div
@@ -4174,7 +4182,7 @@
                     open
                   </div>
                 </div>
-              </div> -->
+              </div>
               <div class="section resetSettings">
                 <h1>reset settings</h1>
                 <div class="text">
